made it int range instead of float

+++ Arduino/ADAPT_test/ADAPT_test.ino
@@ -165,7 +165,7 @@ void setup()
 
 void loop()
 {
-  float range;
+  int range;
   if(off_flag){
     sleep_mode();
   }else{
